requirements.txt
cssjson
Pillow
numpy
PyYAML
requests
